+++ src/main/java/blackjack/domain/player/BettingMoney.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.player;
 
 public class BettingMoney {
