+++ src/main/java/blackjack/domain/player/Players.java
@@ -1,6 +1,6 @@
 package blackjack.domain.player;
 
-import blackjack.domain.CardDeck;
+import blackjack.domain.card.CardDeck;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
